==> scripts/04-biodiversity_statistics.Rmd <==
---
title: "Biodiversity statistics"
output:
  github_document: default
  pdf_document:
    keep_tex: yes
  html_document:
    toc: yes
    toc_depth: 6
    toc_float: yes
editor_options: 
  chunk_output_type: inline
---

# Load libraries

```{r}
library(ggplot2) ## for plotting
library(dplyr) ## for data table manipulation
library(tidyr) ## for data table manipulation
library(readr) ## for reading in tsv files
library(readxl) ## for reading in excel files
library(stringr) ## for data transformation
library(strex) ## for data transformation
library(writexl) ## for excel output
library(purrr) ## for data transformation
library(tidyverse) ## for data transformation
library(BiodiversityR) ## for biodiversity metrics

## for stats
library(vegan)
library(stats)
library(devtools)
#install_github("pmartinezarbizu/pairwiseAdonis/pairwiseAdonis")
library(pairwiseAdonis)
library(lme4) ## for stats
library(car) ## for stats
library(stats) ## for stats
```

# Load data 

```{r}
meta <- read_xlsx("data/metadata/full_metadata.xlsx")
  
data <- read_xlsx("data/results/Relative_abundance_longformat.xlsx") %>% 
  dplyr::select(-Latitude, -Longitude, -`Sequencing Round 1`, -Category, -Common_name) %>%
  spread(Species_name, rel_ab) %>%
  filter(!SampleType == "Blank")

data_matrix <- data %>% dplyr::select(sampleID, `Alopias vulpinus`:`Urophycis sp`) %>%
  column_to_rownames(var = "sampleID")

bottom <- data %>%
  subset(Depth == "Bottom")
bottom_matrix <- bottom %>% dplyr::select(sampleID, `Alopias vulpinus`:`Urophycis sp`) %>%
  column_to_rownames(var = "sampleID")


### no SF 
bottom_noSF <- bottom %>% subset(!Lease_area == "South Fork")

bottom_noSF_matrix <- bottom_noSF %>%
  dplyr::select(sampleID, `Alopias vulpinus`:`Urophycis sp`) %>%
  column_to_rownames(var = "sampleID")

### REVOLUTION WIND control vs inside WEA (bottom sampling)
bottom_REV <- bottom %>% subset(Lease_area == "Revolution Wind")

bottom_REV_matrix <- bottom_REV %>%
  dplyr::select(sampleID, `Alopias vulpinus`:`Urophycis sp`) %>%
  column_to_rownames(var = "sampleID")

### VINEYARD WIND 1 control vs inside WEA (bottom sampling)
bottom_VW <- bottom %>% subset(Lease_area == "Vineyard Wind 1")

bottom_VW_matrix <- bottom_VW %>%
  dplyr::select(sampleID, `Alopias vulpinus`:`Urophycis sp`) %>%
  column_to_rownames(var = "sampleID")
```

## PERMANOVA for beta diversity

```{r}
adonis2(data_matrix ~ Month*Depth*SampleType*Lease_area, data = data, method='eu', permutations = 99)
adonis2(bottom_noSF_matrix ~ Month*SampleType*Lease_area, data = bottom_noSF, method='eu', permutations = 99)

### Revolution Control vs Inside WEA
adonis2(bottom_REV_matrix ~ Month*SampleType, data = bottom_REV, method='eu', permutations = 99)

### VW1 Control vs Inside WEA
adonis2(bottom_VW_matrix ~ Month*SampleType, data = bottom_VW, method='eu', permutations = 99)
```

## NMDS for season

```{r}
vare.mds <- metaMDS(bottom_noSF_matrix, distance = "bray")
data.scores <- as.data.frame(scores(vare.mds)$sites) %>%
  rownames_to_column(var = "sampleID") %>%
  left_join(., meta, by = "sampleID")

mean_data <- data.scores %>%
  mutate(Month = factor(Month, levels = c("July", "August", "September", "October", "November"))) %>%
  group_by(Month, Lease_area) %>%
  summarise(meanNMDS1 = mean(NMDS1), meanNMDS2 = mean(NMDS2)) %>%
  ungroup()  
```

For season 

```{r}
### plot 
data.scores %>%
  mutate(Month = factor(Month, levels = c("July", "August", "September", "October", "November"))) %>%
  
  ggplot(., aes(x=NMDS1, y=NMDS2, shape=Lease_area, linetype = SampleType)) +
  
  geom_point(aes(fill=SampleType), fill_alpha = 0.1, size=2.5) +
  stat_ellipse(level = 0.95, size=1, alpha=0.2) +
  
  theme_bw() + 
  #guides(color = guide_legend(override.aes = list(alpha = 1))) +
  labs(shape = "Lease area", fill = "Sample Type") +
  
  scale_shape_manual(values = c(21,22)) +
  scale_fill_manual(values = c("white", "grey70")) +
  scale_linetype_manual(values = c("solid", "dashed")) +
  
  theme(
    axis.title.y = element_text(margin = margin(t = 0, r = 10, b = 0, l = 0), size=12, face="bold"),
    axis.title.x = element_text(margin = margin(t = 10, r = 0, b = 0, l = 0), size=12, face="bold"),
    axis.text.y = element_text(colour = 'black', size = 10),
    axis.text.x = element_text(colour = 'black', size = 10),
    legend.title = element_text(margin = margin(t = 0, r = 0, b = 5, l = 0), size=10, color="black", face="bold"),
    strip.text.x = element_text(color = "black", face = "bold", size = 12),
    strip.background = element_rect(color= "black", fill = "white")
  ) + 
  
  facet_wrap(~Lease_area)

ggsave("data/results/NMDS_controlvinside.png", width = 6, height = 4)
```


For inside vs outside WEA 

```{r}
### plot 
data.scores %>%
  mutate(Month = factor(Month, levels = c("July", "August", "September", "October", "November"))) %>%
  
  ggplot(., aes(x=NMDS1, y=NMDS2, fill=Month, color=Month, shape=Lease_area)) +
  
  geom_point(alpha=0.2, size=2.5, stroke=0.85) +
  stat_ellipse(level = 0.95, size=1, alpha=0.2) +
  
  geom_path(data=mean_data, aes(x=meanNMDS1, y=meanNMDS2, group = Lease_area), color="black", linetype="solid",
            size=1.15, arrow = arrow(length = unit(0.3, "cm"))) +
  geom_point(data=mean_data, aes(x=meanNMDS1, y=meanNMDS2), color="black", size=4) +

  theme_bw() + 
  guides(fill = "none", color = guide_legend(override.aes = list(alpha = 1))) +
  labs(shape = "Lease area") +

   scale_fill_manual(values = c("#0f4c5c", "#e36414", "#fb8b24", "#9a031e", "#5f0f40")) +
   scale_color_manual(values = c("#0f4c5c", "#e36414", "#fb8b24", "#9a031e", "#5f0f40")) +
  
  scale_shape_manual(values = c(21,22)) +
  theme(
    axis.title.y = element_text(margin = margin(t = 0, r = 10, b = 0, l = 0), size=12, face="bold"),
    axis.title.x = element_text(margin = margin(t = 10, r = 0, b = 0, l = 0), size=12, face="bold"),
    axis.text.y = element_text(colour = 'black', size = 10),
    axis.text.x = element_text(colour = 'black', size = 10),
    legend.title = element_text(margin = margin(t = 0, r = 0, b = 5, l = 0), size=10, color="black", face="bold"),
    strip.text.x = element_text(color = "black", face = "bold", size = 12),
    strip.background = element_rect(color= "black", fill = "white")
  ) + 
  facet_wrap(~Lease_area, ncol=2)

ggsave("data/results/NMDS.png", width = 8.5, height = 4)
```


# Alpha diversity 

```{r}
alpha_df <- read_xlsx("data/results/Relative_abundance_longformat.xlsx") %>%
  group_by(Month, Site, Depth) %>%
  filter(!Category == "unassigned") %>%
  filter(!Category == "Livestock") %>%
  filter(!Category == "Human") %>%
  filter(!Category == "Other") %>%
  filter(rel_ab > 0) %>%
  mutate(richness = n_distinct(Species_name)) %>%
  dplyr::select(Month, Site, Depth, SampleType, Lease_area, richness) %>%
  subset(Depth == "Bottom") %>%
  distinct()

alpha_df %>%
  mutate(Month = factor(Month, levels = c("July", "August", "September", "October", "November"))) %>%
  filter(!is.na(Lease_area)) %>%
  unite(sample_group, Lease_area, SampleType, sep = " ", remove=F) %>%
  
  ggplot(., aes(x=sample_group, y=richness)) +
    theme_bw() +
    xlab("Site") + 
    ylab("Species Richness") + 
    geom_boxplot(aes(fill=sample_group), alpha=0.6, outlier.shape = NA) + 
    geom_jitter(aes(fill=sample_group, shape=Lease_area), size=2, color="black", alpha=0.75, width=0.2) +
  #labs(shape = "Lease area", fill = "Lease area") +
  facet_wrap(~Month, ncol=5) +
  scale_fill_manual(values = c("white", "#264653", "#2a9d8f", "white", "#e9c46a"), name = "Lease Area") +
  scale_shape_manual(values = c(21,24,22)) +
   theme(
    axis.title.y = element_text(margin = margin(t = 0, r = 10, b = 0, l = 0), size=12, face="bold"),
    axis.title.x = element_text(margin = margin(t = 10, r = 0, b = 0, l = 0), size=12, face="bold"),
    axis.text.y = element_text(colour = 'black', size = 10),
    axis.text.x = element_text(angle = 45, size=6, color="grey25", hjust = 1),
    legend.title = element_text(margin = margin(t = 0, r = 0, b = 5, l = 0), size=10, color="black", face="bold"),
    strip.text.x = element_text(color = "black", face = "bold", size = 12),
    strip.background = element_rect(color= "black", fill = "white"),
    legend.position = "none"
  )

ggsave("data/results/species_richness.png", width=10, height=5)
```


```{r}
alpha_df <- alpha_df %>%
  unite(sample_group, Lease_area, SampleType, sep = " ", remove=F) 

aov <- aov(richness ~ Month, data = alpha_df)
Anova(aov, type = "III")
```

### PERMANOVA for bottom surface water

```{r}
surface_bottom_sites <- data %>% filter(!is.na(Depth)) %>% subset(Depth == "Surface") %>%
  dplyr::select(Site) %>% distinct()

data_SvB <- data %>% filter(Site %in% surface_bottom_sites$Site) %>%
  unite(sample_group, Lease_area, SampleType, sep = " ", remove=F)
data_SvB_matrix <- data_SvB %>% dplyr::select(sampleID, `Alopias vulpinus`:`Urophycis sp`) %>%
  column_to_rownames(var = "sampleID")

adonis2(data_SvB_matrix ~ Month*Lease_area*Depth, data = data_SvB, method='eu', permutations = 99)

vare.mds_SvB <- metaMDS(data_SvB_matrix, distance = "bray")
data.scores_SvB <- as.data.frame(scores(vare.mds_SvB)$sites) %>%
  rownames_to_column(var = "sampleID") %>%
  left_join(., meta, by = "sampleID")
```

Plot

```{r}
data.scores_SvB %>%
  unite(sample_group, Lease_area, SampleType, sep = " ", remove=F) %>%
  mutate(Month = factor(Month, levels = c("July", "August", "September", "October", "November"))) %>%
  
  ggplot(., aes(x=NMDS1, y=NMDS2, shape=Lease_area, fill = Depth)) +
  
  geom_point(size=2.5) +
  stat_ellipse(aes(group=Depth), level = 0.95, size=1, alpha=0.2) +
  
  theme_bw() +
  
  scale_fill_manual(values = c("grey90", "grey20"), name = "Sample Type") +
  scale_shape_manual(values = c(24,22)) +
  
  theme(
    axis.title.y = element_text(margin = margin(t = 0, r = 10, b = 0, l = 0), size=12, face="bold"),
    axis.title.x = element_text(margin = margin(t = 10, r = 0, b = 0, l = 0), size=12, face="bold"),
    axis.text.y = element_text(colour = 'black', size = 10),
    axis.text.x = element_text(colour = 'black', size = 10),
    legend.title = element_text(margin = margin(t = 0, r = 0, b = 5, l = 0), size=10, color="black", face="bold"),
    strip.text.x = element_text(color = "black", face = "bold", size = 12),
    strip.background = element_rect(color= "black", fill = "white")
  ) + 
  
  facet_wrap(~Lease_area)
```



```{r}

```
